refactor: split local provider tests
