+++ crates/runner/src/provider/local/tests/mod.rs
@@ -1,0 +1,6 @@
+mod cancellation;
+mod claim;
+mod completion;
+mod discovery;
+mod profiles;
+mod support;
